Support zero amounts and priority with finance update
The finance update command checks whether a parameter has been passed
and only updates the value if it has. With the previous implementation
an existence check is made. This masks out None values, but also masks
out zero values. In order to support zero values, an explicit check for
None is needed.

This changes the check in this way to allow for a zero amount or zero
priority to be passed as a parameter.